Repackage internals, internalize view-only status
This should help prevent unauthorized edits via other plugins for players with view-only access
