+++ internal/v1_21_R1/src/main/java/com/lishid/openinv/internal/v1_21_R1/InternalAccessor.java
@@ -5,9 +5,11 @@ import com.lishid.openinv.internal.IAnySilentContainer;
 import com.lishid.openinv.internal.ISpecialEnderChest;
 import com.lishid.openinv.internal.ISpecialInventory;
 import com.lishid.openinv.internal.ISpecialPlayerInventory;
+import com.lishid.openinv.internal.v1_21_R1.inventory.AnySilentContainer;
 import com.lishid.openinv.internal.v1_21_R1.inventory.OpenEnderChest;
 import com.lishid.openinv.internal.v1_21_R1.inventory.OpenInventory;
 import com.lishid.openinv.internal.v1_21_R1.inventory.Placeholders;
+import com.lishid.openinv.internal.v1_21_R1.player.PlayerManager;
 import com.lishid.openinv.util.lang.LanguageManager;
 import net.minecraft.world.Container;
 import org.bukkit.configuration.ConfigurationSection;
