+++ internal/v1_21_R1/src/main/java/com/lishid/openinv/internal/v1_21_R1/inventory/AnySilentContainer.java
@@ -14,9 +14,10 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.lishid.openinv.internal.v1_21_R1;
+package com.lishid.openinv.internal.v1_21_R1.inventory;
 
 import com.lishid.openinv.internal.IAnySilentContainer;
+import com.lishid.openinv.internal.v1_21_R1.player.PlayerManager;
 import com.lishid.openinv.util.ReflectionHelper;
 import com.lishid.openinv.util.lang.LanguageManager;
 import net.minecraft.core.BlockPos;
@@ -74,18 +75,6 @@ public class AnySilentContainer implements IAnySilentContainer {
         }
     }
 
-    public static @NotNull MenuType<?> getContainers(int inventorySize) {
-        return switch (inventorySize) {
-            case 9 -> MenuType.GENERIC_9x1;
-            case 18 -> MenuType.GENERIC_9x2;
-            case 36 -> MenuType.GENERIC_9x4;
-            case 45 -> MenuType.GENERIC_9x5;
-            case 54 -> MenuType.GENERIC_9x6;
-            // Default to 27-slot inventory
-            default -> MenuType.GENERIC_9x3;
-        };
-    }
-
     @Override
     public boolean activateContainer(
             @NotNull final Player bukkitPlayer,
@@ -114,7 +103,7 @@ public class AnySilentContainer implements IAnySilentContainer {
             PlayerEnderChestContainer enderChest = player.getEnderChestInventory();
             enderChest.setActiveChest(enderChestTile);
             player.openMenu(new SimpleMenuProvider((containerCounter, playerInventory, ignored) -> {
-                MenuType<?> containers = getContainers(enderChest.getContainerSize());
+                MenuType<?> containers = OpenContainerMenu.getContainers(enderChest.getContainerSize());
                 int rows = enderChest.getContainerSize() / 9;
                 return new ChestMenu(containers, containerCounter, playerInventory, enderChest, rows);
             }, Component.translatable("container.enderchest")));
