+++ internal/v1_21_R1/src/main/java/com/lishid/openinv/internal/v1_21_R1/player/PlayerManager.java
@@ -1,20 +1,4 @@
-/*
- * Copyright (C) 2011-2023 lishid. All rights reserved.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, version 3.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.lishid.openinv.internal.v1_21_R1;
+package com.lishid.openinv.internal.v1_21_R1.player;
 
 import com.lishid.openinv.internal.ISpecialInventory;
 import com.lishid.openinv.internal.v1_21_R1.inventory.OpenEnderChest;
